refactor(task): move src/fetch to src/task/fetch
The SEC fetch tasks (SecCachedFetchTask, SecFetchTask, SecFetchJob, SecJobQueue,
SecFetchFileOutputCache) are tasks/task infrastructure, so they belong under
src/task/ with the rest. Move the whole src/fetch directory to src/task/fetch,
rewrite the moved files' outward imports (../config|util|cli → ../../…) and every
importer's path (task-domain siblings ../../fetch → ../fetch; commands/index.ts
../fetch → ../task/fetch). Pure relocation, no behavior change.
