+++ src/task/forms/SecFetchAccessionDocTask.ts
@@ -6,7 +6,7 @@
 
 import { Type } from "typebox";
 import { DataPortSchemaObject, TaskInput } from "workglow";
-import { SecCachedFetchTask } from "../../fetch/SecCachedFetchTask";
+import { SecCachedFetchTask } from "../fetch/SecCachedFetchTask";
 import { TypeSecCik } from "../../sec/submissions/EnititySubmissionSchema";
 
 // NOTE: accession documents are immutable, so we don't need to pass in a date for invalidation
